+++ PlayerQueueService.Api/Telemetry/IMetricsProvider.cs
@@ -8,7 +8,9 @@ public interface IMetricsProvider : IDisposable
     void IncrementPublishSuccess(PlayerEnqueuedEvent playerEvent);
     void IncrementPublishFailure(PlayerEnqueuedEvent playerEvent, string reason);
     void RecordPublishDuration(PlayerEnqueuedEvent playerEvent, double milliseconds);
+    void IncrementPublishRetry(PlayerEnqueuedEvent playerEvent, string queueName);
     void IncrementProcessingRetry(PlayerEnqueuedEvent playerEvent, string queueName);
+    void IncrementProcessed(PlayerEnqueuedEvent playerEvent, string queueName);
     void IncrementConsumeSuccess(PlayerEnqueuedEvent playerEvent, string queueName);
     void IncrementConsumeFailure(PlayerEnqueuedEvent playerEvent, string reason, string queueName);
     void RecordConsumeDuration(PlayerEnqueuedEvent playerEvent, double milliseconds, string queueName);